Update mix predictor with optional axis annotations and support for continuous concepts
